Refactor: App.js 파일포맷 및 위치 변경
- 포맷 : .js -> .jsx
- 디렉토리 위치 : /src -> /src/components
